Factor out PipelineRunStatusLink
