+++ src/components/ImportedApps/PipelineGroupSummary.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 import {
   DescriptionList,
   DescriptionListDescription,
@@ -10,9 +9,7 @@ import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
 
 import { getPipelineGroupSourceNamespace } from 'src/api/pipelineHelpers';
 import { CranePipelineGroup } from 'src/api/types/CranePipeline';
-import { pipelineRunUrl } from 'src/utils/paths';
-import { PipelineRunStatus } from './PipelineRunStatus';
-import { useValidatedNamespace } from 'src/common/hooks/useValidatedNamespace';
+import { PipelineRunStatusLink } from './PipelineRunStatusLink';
 
 interface PipelineGroupSummaryProps {
   pipelineGroup: CranePipelineGroup;
@@ -21,7 +18,6 @@ interface PipelineGroupSummaryProps {
 export const PipelineGroupSummary: React.FunctionComponent<PipelineGroupSummaryProps> = ({
   pipelineGroup,
 }) => {
-  const { namespace } = useValidatedNamespace();
   const latestPipelineRun = pipelineGroup.pipelineRuns.latestNonPending;
   return (
     <DescriptionList
@@ -47,9 +43,7 @@ export const PipelineGroupSummary: React.FunctionComponent<PipelineGroupSummaryP
         <DescriptionListTerm>Last run status</DescriptionListTerm>
         <DescriptionListDescription>
           {latestPipelineRun ? (
-            <Link to={pipelineRunUrl(namespace, latestPipelineRun)}>
-              <PipelineRunStatus pipelineRun={latestPipelineRun} showAction />
-            </Link>
+            <PipelineRunStatusLink pipelineRun={latestPipelineRun} showAction />
           ) : (
             'Not yet attempted'
           )}
